Texture update 0.2
Made some slight changes to the health bar texture to have more of a
glassy look to it, and also tweaked the health bar code for better
sizing.
Added a new texture, the ScrollBorder texture, which is a texture used
for bordered GUI components, so far the only bordered inplementation is
the UIBorderedButton class which is a UIButton modified to work with the
borders.